refact(song.xu):replace base module
